requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-jalali==6.0.1
django-multiupload==0.6.1
idna==3.4
isodate==0.6.1
jdatetime==4.1.1
lxml==4.9.3
Pillow==10.0.1
platformdirs==3.11.0
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
zeep==4.2.1
